move some elements from utils to examples
